requirements.txt
python-dotenv
dotenv
flask
discord.py-self